+++ src/scss/variables.scss
@@ -2,6 +2,7 @@ $primary          : #B5E0D6;
 $primary-dark     : #7EBAAC;
 $secondary        : #2C3E4E;
 $secondary-dark   : #0F1C27;
+$secondary-light  : #334453;
 
 $grey-ultralight  : #F5F5F5;
 $grey-light       : #E3E3E3;
