+++ src/store/rlinks.js
@@ -1,0 +1,79 @@
+/* eslint-disable array-callback-return */
+/* eslint-disable no-return-assign */
+
+export default {
+  state: {
+    rlinks: {},
+    rnodes: {},
+    rlinksHeader: {},
+    rnodesHeader: {},
+    rlineAttributes: [],
+    rnodeAttributes: [],
+  },
+
+  mutations: {
+    loadrLinks (state, payload) {
+      state.rlinks = structuredClone(payload)
+      if (['urn:ogc:def:crs:OGC:1.3:CRS84', 'EPSG:4326'].includes(state.rlinks.crs.properties.name)) {
+        const rlinksHeader = { ...state.rlinks }
+        rlinksHeader.features = []
+        state.rlinksHeader = rlinksHeader
+
+        // limit geometry precision to 6 digit
+        state.rlinks.features.forEach(link => link.geometry.coordinates = link.geometry.coordinates.map(
+          points => points.map(coord => Math.round(Number(coord) * 1000000) / 1000000)))
+        // set all trips visible
+        // this.commit('changeSelectedTrips', state.tripId)
+
+        this.commit('getrLinksProperties')
+        // state.filesAreLoaded.links = true
+      } else { alert('invalid CRS. use CRS84 / EPSG:4326') }
+    },
+
+    loadrNodes (state, payload) {
+      state.rnodes = JSON.parse(JSON.stringify(payload))
+      if (['urn:ogc:def:crs:OGC:1.3:CRS84', 'EPSG:4326'].includes(state.rnodes.crs.properties.name)) {
+        const rnodesHeader = { ...state.rnodes }
+        rnodesHeader.features = []
+        state.rnodesHeader = rnodesHeader
+        // limit geometry precision to 6 digit
+        state.rnodes.features.forEach(node => node.geometry.coordinates = node.geometry.coordinates.map(
+          coord => Math.round(Number(coord) * 1000000) / 1000000))
+
+        this.commit('getrNodesProperties')
+        // state.filesAreLoaded.nodes = true
+      } else { alert('invalid CRS. use CRS84 / EPSG:4326') }
+    },
+    getrLinksProperties (state) {
+      let header = new Set([])
+      state.rlinks.features.forEach(element => {
+        Object.keys(element.properties).forEach(key => header.add(key))
+      })
+      // header.delete('index')
+      // add all default attributes
+      const defaultAttributes = [
+        'index', 'a', 'b']
+      defaultAttributes.forEach(att => header.add(att))
+      header = Array.from(header)
+      state.rlineAttributes = header
+    },
+    getrNodesProperties (state) {
+      let header = new Set([])
+      state.rnodes.features.forEach(element => {
+        Object.keys(element.properties).forEach(key => header.add(key))
+      })
+      // add all default attributes
+      const defaultAttributes = [
+        'index']
+      defaultAttributes.forEach(att => header.add(att))
+      header = Array.from(header)
+      state.rnodeAttributes = header
+    },
+
+  },
+
+  getters: {
+    rlinks: (state) => state.rlinks,
+    rnodes: (state) => state.rnodes,
+  },
+}
